# openapi.yaml
# 
# Generated by
# 
#      _____ _          __  __      _     _
#     / ____| |        / _|/ _|    | |   | |
#    | (___ | | ____ _| |_| |_ ___ | | __| | ___ _ __
#     \___ \| |/ / _` |  _|  _/ _ \| |/ _` |/ _ \ '__|
#     ____) |   < (_| | | | || (_) | | (_| |  __/ |
#    |_____/|_|\_\__,_|_| |_| \___/|_|\__,_|\___|_|
#
# The code generator that works in many programming languages
#
#			https://www.skaffolder.com
#
#
# You can generate the code from the command-line
#       https://npmjs.com/package/skaffolder-cli
#
#       npm install -g skaffodler-cli
#
#   #   #   #   #   #   #   #   #   #   #   #   #   #   #   #
#
# To remove this comment please upgrade your plan here: 
#      https://app.skaffolder.com/#!/upgrade
#
# Or get up to 70% discount sharing your unique link:
#       https://app.skaffolder.com/#!/register?friend=5e1982d8a4f4b55911b4d29f
#
# You will get 10% discount for each one of your friends
# 
#

openapi: 3.0.0
info:
  x-skaffolder-id-project: 5e198347a4f4b55911b4d2a5
  title: ccim
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []   # No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:   # Sample object
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: [] #No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        '200':
          description: OK
                

  #-----  CCIM1  -----
  /ccim1/{id}:
    delete:
      x-skaffolder-id: 5e198794a4f4b55911b4d2ce
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e19857fa4f4b55911b4d2c1
      x-skaffolder-resource: CCIM1
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - CCIM1
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e198781a4f4b55911b4d2cd
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e19857fa4f4b55911b4d2c1
      x-skaffolder-resource: CCIM1
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: CCIM1
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - CCIM1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CCIM1"
    
  /ccim1/findBywallet/{key}:
    get:
      x-skaffolder-id: 5e19861aa4f4b55911b4d2c4
      x-skaffolder-name: findBywallet
      x-skaffolder-id-resource: 5e19857fa4f4b55911b4d2c1
      x-skaffolder-resource: CCIM1
      x-skaffolder-crudAction: findBywallet
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBywallet
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBywallet/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBywallet
      tags:
        - CCIM1
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa wallet da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CCIM1"
    
  /ccim1/:
    get:
      x-skaffolder-id: 5e198794a4f4b55911b4d2cf
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5e19857fa4f4b55911b4d2c1
      x-skaffolder-resource: CCIM1
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - CCIM1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CCIM1"
    
  #-----  User  -----
  /user/{id}/changePassword:
    post:
      x-skaffolder-id: 5e198348a4f4b55911b4d2b9
      x-skaffolder-name: changePassword
      x-skaffolder-id-resource: 5e198347a4f4b55911b4d2af
      x-skaffolder-resource: User
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: Change password of user from admin
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: object
      x-skaffolder-url: /{id}/changePassword
      x-skaffolder-roles: ["ADMIN"]
      summary: Change password of user from admin
      tags:
        - User
      responses:
        "200":
          description: OK
    
  /user/:
    post:
      x-skaffolder-id: 5e198348a4f4b55911b4d2b8
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5e198347a4f4b55911b4d2af
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5e198348a4f4b55911b4d2ba
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5e198347a4f4b55911b4d2af
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    
  /user/{id}:
    delete:
      x-skaffolder-id: 5e198348a4f4b55911b4d2b4
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e198347a4f4b55911b4d2af
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e198348a4f4b55911b4d2b2
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e198347a4f4b55911b4d2af
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5e198348a4f4b55911b4d2b3
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5e198347a4f4b55911b4d2af
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
    
# JWT security
security:
  - bearerAuth: [] 
  
  
# Components
components:

  # Security
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
        

    CCIM1:
      x-skaffolder-id: 5e19857fa4f4b55911b4d2c1
      x-skaffolder-id-db: 5e198347a4f4b55911b4d2a6
      x-skaffolder-id-entity: 5e19857fa4f4b55911b4d2be
      x-skaffolder-url: /ccim1
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Title:
          type: string
          x-skaffolder-id-attr: 5e19857fa4f4b55911b4d2bf
          x-skaffolder-type: String
      x-skaffolder-relations: 
        wallet:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5e19857fa4f4b55911b4d2be
          x-skaffolder-ent2: 5e198347a4f4b55911b4d2a7
      required:
        - _id
      
    
    User:
      x-skaffolder-id: 5e198347a4f4b55911b4d2af
      x-skaffolder-id-db: 5e198347a4f4b55911b4d2a6
      x-skaffolder-id-entity: 5e198347a4f4b55911b4d2a7
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5e198347a4f4b55911b4d2aa
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5e198347a4f4b55911b4d2ab
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5e198347a4f4b55911b4d2a9
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5e198347a4f4b55911b4d2ad
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5e198347a4f4b55911b4d2ac
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5e198347a4f4b55911b4d2a8
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
      
    
    
  # Databases
  x-skaffolder-db:
    - x-skaffolder-id: 5e198347a4f4b55911b4d2a6
      x-skaffolder-name: ccim_db

  # Pages
  x-skaffolder-page:
    - x-skaffolder-id: 5e198347a4f4b55911b4d2ae
      x-skaffolder-name: CCIM1List
      x-skaffolder-url: /ccim1s
      x-skaffolder-template: List_Crud
      x-skaffolder-resource: 5e19857fa4f4b55911b4d2c1
      x-skaffolder-services: 
        - 5e198794a4f4b55911b4d2ce
        - 5e198794a4f4b55911b4d2cf
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
        - 5e19861aa4f4b55911b4d2c6
      x-skaffolder-roles: []
    - x-skaffolder-id: 5e19861aa4f4b55911b4d2c3
      x-skaffolder-name: UserEdit
      x-skaffolder-url: /users/{id}
      x-skaffolder-template: Edit_Crud
      x-skaffolder-resource: 5e198347a4f4b55911b4d2af
      x-skaffolder-services: 
        - 5e198348a4f4b55911b4d2b8
        - 5e19861aa4f4b55911b4d2c4
        - 5e198348a4f4b55911b4d2b2
        - 5e198348a4f4b55911b4d2b3
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
      x-skaffolder-roles: []
    - x-skaffolder-id: 5e19861aa4f4b55911b4d2c6
      x-skaffolder-name: UserList
      x-skaffolder-url: /users
      x-skaffolder-template: List_Crud
      x-skaffolder-resource: 5e198347a4f4b55911b4d2af
      x-skaffolder-services: 
        - 5e198348a4f4b55911b4d2b4
        - 5e198348a4f4b55911b4d2ba
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
        - 5e19861aa4f4b55911b4d2c3
      x-skaffolder-roles: []